chore(redis-cluster): Update Redis Cluster configuration to support multiple hosts
This commit updates the Redis Cluster configuration in the `redis-cluster-provider.ts` file to support multiple hosts. It adds a new `hosts` property to the `IRedisClusterProviderConfig` interface and modifies the `getRedisClusterProviderConfig` function to handle multiple hosts. The `getRedisCluster` function is also updated to use the new `hosts` property when creating the Redis Cluster instance.

```